refactor workflows to split
