+++ .github/workflows/main.yml
@@ -88,8 +88,6 @@ jobs:
         with:
           go-version-file: './go.mod'
           check-latest: true
-      - name: run e2e test
-        run: ./pkg/test/e2e/e2e_test.sh
-      - name: run identity monitor test
-        run: ./pkg/test/identity_workflow/identity_workflow_e2e_test.sh
+      - name: run Rekor end-to-end test
+        run: ./pkg/test/rekor_e2e/rekor_monitor_e2e_test.sh
